requirements.txt
streamlit
pyyaml
anthropic
netmiko
